libautomotive::types

Trait Port

pub trait Port: Send + Sync {
    // Required methods
    fn send(&mut self, frame: &Frame) -> Result<()>;
    fn receive(&mut self) -> Result<Frame>;
    fn set_timeout(&mut self, timeout_ms: u32) -> Result<()>;
}

Hardware abstraction trait for CAN interfaces.

This trait must be implemented by platform-specific code to provide the actual hardware communication capabilities.

Required Methods

fn send(&mut self, frame: &Frame) -> Result<()>

Sends a frame through the CAN interface.

fn receive(&mut self) -> Result<Frame>

Receives a frame from the CAN interface.

This method will block until a frame is received or a timeout occurs.

fn set_timeout(&mut self, timeout_ms: u32) -> Result<()>

Sets the timeout for receive operations.

Parameters

timeout_ms - Timeout in milliseconds. A value of 0 means no timeout.